Handwritten Character Recognition Using Deep Learning
The objective of this project is to develop and evaluate deep learning models for recognizing handwritten characters using the EMNIST Balanced dataset. This project involves creating and fine-tuning two types of neural network architectures: Multi-Layer Perceptron (MLP) and Convolutional Neural Network (CNN). The goal is to identify the most effective model and hyperparameter configurations for achieving high accuracy in character recognition tasks.